% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_check.r
\name{cam_check}
\alias{cam_check}
\title{Check camera trap photos & directories}
\usage{
cam_check(path)
}
\arguments{
\item{path}{character; base path for all photos}
}
\value{
\code{cam_check} returns an S3 object of class "cam_check", which is
  a named list with each element corresponding to results of a different
  check. Call \code{print} to see a summary of the number of errors, or
  potential errors, found in each category.
}
\description{
Check to ensure that a directory of camera trap photos is correctly organized
into subfolders.
}
\details{
\code{\link{cam_process}} requires that camera trap photos are organized into
a strict, hierarchical directory structure. \code{cam_check} ensure that this
structure is adhered to and should always be run prior to
\code{\link{cam_process}}.

Every photo should be placed into a series of directories such as
\code{site/camera/species/number/photo.jpg} where:

\enumerate{
 \item \code{site}: names of the location where cameras were placed
 \item \code{camera}: unique identifiers for each camera
 \item \code{species}: name of species identified in photo
 \item \code{number}: number (e.g. 3) of individuals in photo. If the number
     of individuals is unknown use "x" for the directory name.
 \item \code{photo.jpg}: photos are placed at the end of this chain of
     directories, the actual name of the photo is irrelevant, but it must
     have an EXIF datetime stamp
}

\code{cam_check} returns a named list with eight elements (each a character
vector) corresponding to the eight checks that it performs:

\itemize{
 \item \code{ignore}: photos ignored because they're it directories
     named "ignore".
 \item \code{directory_problem}: photos that are not correctly nested within
     subdirectory structure.
 \item \code{missing_date}: photos that have no associated EXIF date
 \item \code{name_problem}: directories with spaces or special characters
     in the name; only letters, numbers, and _ or - should occur.
 \item \code{count_problem}: count directories that are not either integers
     or "x".
 \item \code{site}: alphabetic list of unique sites; check for typos.
 \item \code{camera}: alphabetic list of unique cameras; check for typos.
 \item \code{species}: alphabetic list of unique species; check for typos.
}
}
\examples{
messy_path <- system.file("extdata", "messy", package = "camtrapr")
checks <- cam_check(messy_path)
checks
checks$directory_problem
checks$species
}

