% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_process.r
\name{cam_process}
\alias{cam_process}
\title{Process Camera Trap Photos}
\usage{
cam_process(path, strict = TRUE)
}
\arguments{
\item{path}{character; base path for all photos}

\item{strict}{logical}
}
\value{
Data frame of processed camera trapping photo data.
}
\description{
Process Camera Trap Photos
}
\examples{
photo_path <- system.file("extdata", "example_photos", package = "camtrapr")
cam_process(photo_path)
}

