\name{camtrapR-defunct}
\alias{camtrapR-defunct}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Consistency check on species image identification
}
\description{

The following functions are now defunct (no longer available):

}

\details{

\

}


\author{
Juergen Niedballa
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
