% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapR-package.R
\docType{package}
\name{camtrapR-package}
\alias{camtrapR-package}
\alias{camtrapR}
\title{Overview of the functions in the camtrapR package}
\description{
This package provides a streamlined workflow for processing data generated
in camera trap-based wildlife studies and prepares input for further
analyses, particularly in occupancy and spatial capture-recapture
frameworks. It suggests a simple data structure and provides functions for
managing digital camera trap photographs (and videos), generating record
tables, maps of species richness and species detections and species activity
diagrams. It further helps prepare subsequent analyses by creating
detection/non-detection matrices for occupancy analyses, e.g. in the
\pkg{unmarked} or \pkg{ubms} packages, and \code{capthist} objects for
spatial capture-recapture analyses in the \pkg{secr} package. In addition,
basic survey statistics are computed. The functions build on one another in
a logical sequence. The only manual input needed is species (and individual)
identification, which is achieved by moving images into species directories
or by tagging images in image management software. Besides, a table holding
basic information about camera trap station IDs, locations and trapping
periods must be created in spreadsheet software.
}
\details{
Image metadata (such as date and time or user-assigned tags) are extracted
from the images using Phil Harvey's ExifTool (available from
\url{https://exiftool.org/}) and the information is stored in a record
table. An adjustable criterion for temporal independence of records can be
applied. Maps of species presence and species richness can be generated.
Several functions are available for plotting single- and two-species
activity patterns. Information about the camera-specific trapping periods
(and periods of malfunction) are summarized into a matrix about camera trap
operability. These, together with the record table, are used to generate
species detection histories for occupancy and spatial capture-recapture
analyses. The user has considerable freedom in generating the detection
histories; sampling occasion length, beginning date and and occasion start
times are adjustable. In addition, trapping effort (i.e. active trap nights
per station and occasion) can be computed for use as a covariate / offset on
detection probability.
}
\section{User support}{
 The camtrapR Google group is an online support and
help forum for camtrapR users. You can find it here:
\url{https://groups.google.com/forum/#!forum/camtrapr}.
}

\references{
Niedballa, J., Sollmann, R., Courtiol, A., Wilting, A. (2016): camtrapR: an
R package for efficient camera trap data management. Methods in Ecology and
Evolution, 7(12).
\url{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.12600}
\cr

camtrapR Google Group \url{https://groups.google.com/forum/#!forum/camtrapr}
\cr

Lemon, J. (2006) Plotrix: a package in the red light district of R. R-News,
6(4): 8-12. \cr

Mike Meredith and Martin Ridout (2018). overlap: Estimates of coefficient of
overlapping for animal activity patterns. R package version 0.3.2.
\url{https://CRAN.R-project.org/package=overlap} \cr

Phil Harvey's ExifTool \url{https://exiftool.org/} \cr
}
\seealso{
\pkg{overlap} \pkg{unmarked} \pkg{ubms} \pkg{secr} \pkg{wiqid}
}
\author{
Juergen Niedballa

Maintainer:Juergen Niedballa <camtrapr@gmail.com>
}
\keyword{internal}
