% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variousOtherHelperFunctions.R
\name{.check_memory_usage}
\alias{.check_memory_usage}
\title{validate memory before allocation}
\usage{
.check_memory_usage(ram_usage_estimate)
}
\arguments{
\item{ram_usage_estimate}{Estimated size of the object in GB}
}
\description{
validate memory before allocation
}
