% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordTable.R
\name{dplyr_reconstruct.records}
\alias{dplyr_reconstruct.records}
\title{Wrapper to discard class in dplyr}
\usage{
\method{dplyr_reconstruct}{records}(data, template)
}
\arguments{
\item{data}{A tibble. We use tibbles because they avoid some inconsistent
subset-assignment use cases.}

\item{template}{Template data frame to use for restoring attributes.}
}
\description{
Wrapper to discard class in dplyr
}
