% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionHistory.R
\name{print.detHist}
\alias{print.detHist}
\title{Printing method for detection history}
\usage{
\method{print}{detHist}(x, nRows = 4, nCols = 3, digits = 3, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{nRows}{the number of first and last rows to display}

\item{nCols}{the number of first and last columns to display}

\item{digits}{the number of digits to use within cells}

\item{...}{further arguments passed to or from other methods}
}
\description{
Printing method for detection history
}
\keyword{internal}
