% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordTable.R
\name{print.records}
\alias{print.records}
\title{Printing method for record tables}
\usage{
\method{print}{records}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\description{
Printing method for record tables
}
\keyword{internal}
