% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapR-package.R, R/data.R
\docType{data}
\name{recordTableSample}
\alias{recordTableSample}
\title{Sample species record table from camera trap images}
\format{
A data frame with 39 rows and 11 variables

A data frame with 39 rows and 12 variables
}
\usage{
data(recordTableSample)
}
\description{
Sample species record table from camera trap images generated from the
sample images in the package with the function \code{\link{recordTable}} .

Sample species record table from camera trap images generated from the
sample images in the package with the function \code{\link{recordTable}}.
}
\details{
The variables are as follows:

\itemize{ \item Station.  Camera trap station ID \item Species.  Species ID
\item DateTimeOriginal.  Date and time as extracted from image \item Date.
record date \item Time.  record time of day \item delta.time.secs.  time
difference to first species record at a station (seconds) \item
delta.time.mins.  time difference to first species record at a station
(minutes) \item delta.time.hours.  time difference to first species record
at a station (hours) \item delta.time.days.  time difference to first
species record at a station (days) \item Directory.  Image directory \item
FileName.  image filename }

The variables are as follows:
\tabular{ll}{
\code{Station} \tab camera trap station ID \cr
\code{Species} \tab species ID \cr
\code{DateTimeOriginal} \tab date and time as extracted from image \cr
\code{Date} \tab record date \cr
\code{Time} \tab record time of day \cr
\code{delta.time.secs} \tab time difference to first species record at a station (seconds) \cr
\code{delta.time.mins} \tab time difference to first species record at a station (minutes) \cr
\code{delta.time.hours} \tab time difference to first species record at a station (hours) \cr
\code{delta.time.days} \tab time difference to first species record at a station (days) \cr
\code{Directory} \tab image directory \cr
\code{FileName} \tab image filename \cr
\code{n_images} \tab the number of images \cr
}
}
\keyword{datasets}
