% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityModel.R
\name{summary,commOccu-method}
\alias{summary,commOccu-method}
\title{Summarize community occupancy model}
\usage{
\S4method{summary}{commOccu}(object, ...)
}
\arguments{
\item{object}{\code{commOccu} object}

\item{...}{currently ignored}
}
\value{
Model summary printed to console
}
\description{
Gives an overview of the number of species, stations and occasions in a \code{commOccu} object. Also returns covariates.
}
\details{
The summary method is very basic and still work in progress.
}
