% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameraOperation.R
\name{summary.camOp}
\alias{summary.camOp}
\title{Summary method for camera trap station operation matrices}
\usage{
\method{summary}{camOp}(object, nStationsMax = 10, ...)
}
\arguments{
\item{object}{an object of class \code{camOp}}

\item{nStationsMax}{integer. Maximum number of rows to show in the
station-level table.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for camera trap station operation matrices
}
\keyword{Summary}
\keyword{camera}
\keyword{for}
\keyword{internal}
\keyword{matrices}
\keyword{method}
\keyword{operation}
\keyword{station}
\keyword{trap}
