% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordTable.R
\name{summary.records}
\alias{summary.records}
\title{Summary method for record tables}
\usage{
\method{summary}{records}(object, nSpeciesMax = 6, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for record tables
}
\keyword{internal}
