% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_check.r
\name{cam_check}
\alias{cam_check}
\title{Check camera trap photos & directories}
\usage{
cam_check(path)
}
\arguments{
\item{path}{character; base path for all photos}
}
\value{
\code{cam_check} returns an S3 object of class "cam_check", which is
       a named list with each element corresponding to results of a different
       check. Call \code{print} to see a summary of the number of errors, or
       potential errors, found in each category.
}
\description{
Check camera trap photos & directories
}
\examples{
messy_path <- system.file("extdata", "messy", package = "camtrapr")
checks <- cam_check(messy_path)
checks
checks$directory_problem
checks$species
}

