% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_process.r
\name{cam_process}
\alias{cam_process}
\title{Process camera trap photos}
\usage{
cam_process(path, clean_names = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{character; base path for all photos}

\item{clean_names}{logical idicating whether to clean up site, camera, and
species names. If \code{TRUE}, leading and trailing whitespace is
trimmed, any non-alphanumeric character will be converted to an
underscore, then groups of more than one underscore will be collapsed
to one. WARNING: if names are cleaned it is possible that previously
unique site, camera, or species names will no longer be unique.}

\item{verbose}{logical indicating whether to print messages highlighting
potential issues including directories than have been ignore, invalid
directory structures, and count directories not corresponding to
integers.}
}
\value{
Data frame of processed camera trapping photo data.
}
\description{
Process camera trap photos
}
\examples{
photo_path <- system.file("extdata", "example_photos", package = "camtrapr")
cam_data <- cam_process(photo_path)
dplyr::glimpse(cam_data)
messy_path <- system.file("extdata", "messy", package = "camtrapr")
messy_data <- cam_process(messy_path)
dplyr::glimpse(messy_data)
}

