% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityModel.R
\encoding{UTF-8}
\name{communityModel}
\alias{communityModel}
\title{Create a community (multi-species) occupancy model for JAGS or Nimble}
\usage{
communityModel(
  data_list,
  occuCovs = list(fixed = NULL, ranef = NULL),
  detCovs = list(fixed = NULL, ranef = NULL),
  detCovsObservation = list(fixed = NULL, ranef = NULL),
  effortCov = "effort",
  intercepts = list(det = "fixed", occu = "fixed"),
  richnessCategories = NULL,
  augmentation = NULL,
  modelFile = NULL,
  nimble = FALSE
)
}
\arguments{
\item{data_list}{list. Contains 3 slots: ylist, siteCovs, obsCovs. ylist ist a list of detection histories (can be named), e.g. from \code{\link{detectionHistory}}. siteCovs is a data.frame with site covariates (optional). obsCovs is a list of site-occasion level covariates (e.g. site-occasion-specific effort, which is also returned by \code{\link{detectionHistory}}.}

\item{occuCovs}{list. Up to 2 items named "fixed" and/or "ranef". Specifies fixed or random effects of covariates on occupancy probability (continuous or categorical covariates)}

\item{detCovs}{list. Up to 2 items named "fixed" and/or "ranef". Specifies fixed or random effects of covariates on detection probability (continuous or categorical covariates)}

\item{detCovsObservation}{list. Up to 2 items named "fixed" and/or "ranef". Specifies  fixed or random effects of observation-level covariates on detection probability  (continuous or categorical covariates - categorical must be coded as character matrix)}

\item{effortCov}{character. Name of list item in \code{data_list$obsCovs} which contains effort. This does not include effort as a covariate on detection probability, but only uses NA / not NA information to create binary effort and ensure detection probabilities p are 0 when there was no effort (p will be 0 whereever \code{effortCov} is NA).}

\item{intercepts}{list. For detection and occupancy probability intercepts, are they "fixed" (= constant across species) or "ranef" (= random effect of species on intercept)?}

\item{richnessCategories}{character. Name of categorical covariate in \code{data_list$siteCovs} for which to calculate separate richness estimates (optional). Can be useful to obtain separate richness estimates for different areas.}

\item{augmentation}{If NULL, no data augmentation (only use species in \code{data_list$ylist}), otherwise named list or vector with total number of (potential) species. Names: "knownmax" or "full". Example: \code{augmentation = c(knownmax = 30)} or \code{augmentation = c(full = 30)}}

\item{modelFile}{character. Text file name to save model to}

\item{nimble}{logical. If TRUE, model code will be for Nimble (incompatible with JAGS). If FALSE, model code is for JAGS.}
}
\value{
\code{commOccu} object. It is an S4 class with the following slots:

  \item{modelText}{JAGS model code as a character vector (made up of code chunks, use cat() to print)}
  \item{params}{Parameters to monitor in the model runs}
  \item{inits_fun}{Function to create start values for the MCMC chains. It being a function ensures different values in each chain}
  \item{data}{List with data needed to run the model (detection & effort matrices, site covariates, number of species / stations / occasions)}
  \item{input}{Input data_list (unchanged)}
  \item{nimble}{logical indicator for whether it is a Nimble model}
  \item{modelFile}{Path of the text file containing the model code}
  
See \code{\link{commOccu-class}}
}
\description{
Flexibly creates complete code and input data for community occupancy models for in JAGS amd Nimble, and automatically sets initial values and parameters to monitor. Supports fixed and random effects of covariates on detection and occupancy probabilities, using both continuous and categorical covariates (both site and site-occasion covariates). 

Optionally includes data augmentation (fully open community, or up to known maximum number of species, or no data augmentation). Allows combination of all these parameters for fast and flexible customization of community occupancy models.

Incidentally, the function can also be used to create  model code and input for single-species single-season occupancy models (it is the special case with only one species). Model will run slower than proper single-species model JAGS code due to the additional species loop, but it is possible.

The function returns several derived quantities, e.g. species richness, Bayesian p-values (overall and by species), Freeman-Tukey residuals for actual and simulate data (by station and total). If doing data augmentation, metacommunity size and number of unseen species are returned also.
}
\details{
For examples of implementation, see Vignette XXX.

Fixed effects of covariates are constant across species, whereas random effect covariates differ between species. Fixed and random effects are allowd for station-level detection and occupancy covariates (a.k.a. site covariates), and also station-occasion level covariates (a.k.a. observation covariates). 

By default, random effects will be by species. It is however possible to use categorical site covariates for grouping (continuous|categorical).
Furthermore, is is possible to use use nested random effects of species and another categorical site covariate (so that there is a random effect of species and an additional random effect of a categorical covariate within each species).


Derived quantities returned by the model are:
  
  \tabular{ll}{
    \code{Bpvalue} \tab Bayesian p-value (overall) \cr
    \code{Bpvalue_species} \tab Bayesian p-value (by species) \cr
    
    
    \code{Nspecies} \tab Species richness (only in JAGS model)\cr
    \code{Nspecies_Covariate} \tab Species richness by categorical covariate (when using \code{richnessCategories}, only in JAGS model) \cr
    
    
    \code{R2} \tab sum of Freeman-Tukey residuals of observed data within each species \cr
    \code{new.R2} \tab sum of Freeman-Tukey residuals of simulated data within each species \cr
    \code{R3} \tab Total sum of Freeman-Tukey residuals of observed data \cr 
    \code{new.R3} \tab Total sum of Freeman-Tukey residuals of simulated data \cr
    \emph{\code{Ntotal}} \tab Total metacommunity size (= observed species + n0) \cr
    \emph{\code{n0}} \tab Number of unseen species in metacommunity \cr
    \emph{\code{omega}} \tab Data augmentation parameter \cr
    \emph{\code{w}} \tab Metacommunity membership indicator for each species
  }

Quantities in \emph{italic} at the bottom are only returned in full data augmentation. \code{Nspecies} and \code{Nspecies_Covariate} are only returned in JAGS models (because Nimble models don't explicitly return latent occupancy status z).
}
\section{Parameter naming convention}{
 
The parameter names are assembled from building blocks. The nomenclature  is as follows:    
  \tabular{lll}{
    \bold{Type}     \tab \bold{Name}         \tab  \bold{Description}  \cr
    Submodel        \tab \bold{\code{alpha}} \tab  detection submodel  \cr
    Submodel        \tab \bold{\code{beta}}  \tab  occupancy submode \cr
    Intercept       \tab \bold{\code{0}}     \tab  denotes the intercepts (alpha0, beta0)  \cr
    Effect type     \tab \bold{\code{fixed}} \tab  fixed effects (constant across species)  \cr
    Effect type     \tab \bold{\code{ranef}} \tab  random effects (of species and/or other categorical covariates)  \cr
    Covariate type  \tab \bold{\code{cont}}  \tab  continuous covariates \cr
    Covariate type  \tab \bold{\code{categ}} \tab  categorical covariates \cr
    Hyperparameter  \tab \bold{\code{mean}}  \tab  mean of random effect \cr
    Hyperparameter  \tab \bold{\code{sigma}} \tab  standard deviation of random effect \cr
    Hyperparameter  \tab \bold{\code{tau}}   \tab  precision of random effect (used internally, not returned)
  }


For example, a fixed intercept of occupancy (constant across species) is \bold{\code{beta0}}.


An intercept with a random effect of species is: 
  
\bold{\code{beta0.mean}} community mean of the occupancy probability intercept (= logit(occupancy probability) at covariate values of 0).

\bold{\code{beta0.sigma}} standard deviation of the estimated occupancy probability intercept.

\bold{\code{beta0[1]}} occupancy intercept of species 1 (cont. for other species). 


For effects of site covariates, the patters is:
  
\code{submodel.effectType.covariateType.CovariateName.hyperparameter}

For example:
  
\bold{\code{beta.ranef.cont.habitat.mean}} is the mean community effect of the continuous site covariate 'habitat' on occupancy probability

\bold{\code{beta.ranef.cont.habitat[1]}} is the effect of continuous site covariate 'habitat' on occupancy probability of species 1

Site-occasion covariates are denoted by ".obs" after the submodel, e.g.: 
  
\bold{\code{alpha.obs.fixed.cont.effort}} is the fixed effect (constant across species) of the continuous observation-level covariate 'effort' on detection probability
}

\examples{

\dontrun{

data("camtraps")

# create camera operation matrix
camop_no_problem <- cameraOperation(CTtable      = camtraps,
                                    stationCol   = "Station",
                                    setupCol     = "Setup_date",
                                    retrievalCol = "Retrieval_date",
                                    hasProblems  = FALSE,
                                    dateFormat   = "dmy"
)

data("recordTableSample")

# list of detection histories
DetHist_list <- lapply(unique(recordTableSample$Species), FUN = function(x) {
  detectionHistory(
    recordTable         = recordTableSample,
    camOp                = camop_no_problem,
    stationCol           = "Station",
    speciesCol           = "Species",
    recordDateTimeCol    = "DateTimeOriginal",
    species              = x,
    occasionLength       = 7,
    day1                 = "station",
    datesAsOccasionNames = FALSE,
    includeEffort        = TRUE,
    scaleEffort          = TRUE,
    timeZone             = "Asia/Kuala_Lumpur"
  )}
)

names(DetHist_list) <- unique(recordTableSample$Species)

ylist <- lapply(DetHist_list, FUN = function(x) x$detection_history)

# some fake covariates for demonstration
sitecovs <- camtraps[, c(1:3)]
sitecovs$elevation <- c(300, 500, 600)   
sitecovs[, c(2:4)] <- scale(sitecovs[,-1])   # scale numeric covariates


# bundle input data for communityModel
data_list <- list(ylist = ylist,
                  siteCovs = sitecovs,
                  obsCovs = list(effort = DetHist_list[[1]]$effort))


# create community model for JAGS
modelfile1 <- tempfile(fileext = ".txt")
mod.jags <- communityModel(data_list,
                           occuCovs = list(fixed = "utm_y", ranef = "elevation"),
                           detCovsObservation = list(fixed = "effort"),
                           intercepts = list(det = "ranef", occu = "ranef"),
                           modelFile = modelfile1)

summary(mod.jags)

# fit in JAGS
fit.jags <- fit(mod.jags,
                n.iter = 1000,
                n.burnin = 500,
                chains = 3)   
summary(fit.jags)

# response curves (= marginal effect plots)
plot_effects(mod.jags, 
             fit.jags, 
             submodel = "state")
plot_effects(mod.jags, 
             fit.jags, 
             submodel = "det")
             
# effect sizes plot
plot_coef(mod.jags, 
          fit.jags, 
          submodel = "state")
plot_coef(mod.jags, 
          fit.jags, 
          submodel = "det")              

# create community model for Nimble
modelfile2 <- tempfile(fileext = ".txt")
mod.nimble <- communityModel(data_list,
                             occuCovs = list(fixed = "utm_x", ranef = "utm_y"),
                             detCovsObservation = list(fixed = "effort"),
                             intercepts = list(det = "ranef", occu = "ranef"),
                             modelFile = modelfile2, 
                             nimble = TRUE)      # set nimble = TRUE

# load nimbleEcology package 
# currently necessary to do explicitly, to avoid additional package dependencies
require(nimbleEcology)

# fit uncompiled model in Nimble
fit.nimble.uncomp <- fit(mod.nimble, 
                         n.iter = 10, 
                         chains = 1)

# fit compiled model in Nimble
fit.nimble.comp <- fit(mod.nimble, 
                       n.iter = 5000, 
                       n.burnin = 2500,
                       chains = 3, 
                       compile = TRUE)

# parameter summary statistics
summary(fit.nimble.comp)


# response curves (= marginal effect plots)
plot_effects(mod.nimble, 
             fit.nimble.comp, 
             submodel = "state")
plot_effects(mod.nimble, 
             fit.nimble.comp, 
             submodel = "det")

#' # effect sizes plot
plot_coef(mod.nimble, 
          fit.nimble.comp, 
          submodel = "state")
plot_coef(mod.nimble, 
          fit.nimble.comp, 
          submodel = "det")   

# traceplots
plot(fit.nimble.comp)


}

}
\references{
Kéry, M., and J. A. Royle. "Applied hierarchical modelling in ecology - Modeling distribution, abundance and species richness using R and BUGS." Volume 1: Prelude and Static Models. Elsevier/Academic Press, 2016.
}
\author{
Juergen Niedballa
}
