% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCovariates.R
\name{createCovariates}
\alias{createCovariates}
\title{Extract covariate values from spatial rasters and prepare rasters for spatial predictions}
\usage{
createCovariates(
  CTtable,
  directory,
  filenames,
  buffer_ct = 0,
  bilinear = FALSE,
  buffer_aoi = 1000,
  raster_template,
  resolution,
  append = TRUE,
  formats = ".tif"
)
}
\arguments{
\item{CTtable}{sf object as defined by the \code{sf} package. Essentially a camera trap data frame with spatial information.}

\item{directory}{character. The directory containing the covariate rasters.}

\item{filenames}{character (optionally named). A vector of file paths of covariate rasters. If it is named the covariates will be named according to the names. If unnamed the file names will be used as covariate names.}

\item{buffer_ct}{numeric. A value (in meters) by which to buffer the point locations in \code{CTtable} for extraction of covariate values.}

\item{bilinear}{logical. If \code{TRUE}, extract covariate values with bilinear interpolation (nearest 4 raster cells). If \code{FALSE}, extract value at the cell the point falls in. Only relevant if \code{buffer_ct} is 0.}

\item{buffer_aoi}{numeric. A value (in meters) by which to buffer the overall camera trapping grid to ensure that prediction rasters are larger than the camera trapping grid.}

\item{raster_template}{SparRaster. A SparRaster (as defined in the \code{terra} package) to use as template for the creation of prediction rasters.}

\item{resolution}{numeric. Spatial resolution of prediction rasters. Ignored if \code{raster_template} is provided.}

\item{append}{logical. If \code{TRUE}, add the extracted covariates to the existing \code{CTtable}. If \code{FALSE}, return only the extracted covariate values without the existing \code{CTtable}.}

\item{formats}{character. Possible file formats for raster data. Defaults to \code{.tif} files.}
}
\value{
\describe{
  \item{CTtable}{An \code{sf} object representing the camera trap data frame with extracted covariate values.}
  \item{prediction_raster}{A \code{SpatRaster} object containing prediction rasters.}
}
}
\description{
This function extracts covariate values from spatial raster data 
 (e.g. for use in modelling) and prepares these covariates for use in 
 spatial predictions. 
 
 It accepts a camera trap table containing spatial information, 
 along with either a directory containing covariate raster files or a character 
 vector specifying the file paths of these covariate rasters. 
 
 Additionally, users can provide parameters to control how covariates are
 extracted, and how they are aggregated to prediction rasters. 
 
 The function generates prediction rasters based on a provided template or 
 creates one automatically if no template is provided.
 
 The function returns a list containing the camera trap dataset with 
 extracted covariate values (e.g for use in occupancy modelling), and 
 prediction rasters ready for spatial modeling.
}
\details{
The input camera trap table must be an \code{\link[sf]{sf}} object (a data 
frame with a geometry column specifying the spatial information). For details 
on how to convert an exising camera trap table to sf, see the Examples below.
The input rasters can be in different coordinate systems. During covariate 
extraction the CTtable is projected to each raster's coordinate system 
individually. For the prediction raster all input rasters are either 
resampled or reprojected to a consistent coordinate system.
}
\examples{

\dontrun{
# load camera trap table
data(camtraps)

# create sf object
camtraps_sf <- st_as_sf(camtraps, 
                        coords = c("utm_x", "utm_y"), 
                        crs = 32650)

# extract covariates (with 100m buffer around cameras)
camtraps_sf_cov <- createCovariates(camtraps_sf, 
"path/to/covariate_rasters", 
buffer_ct = 100, 
buffer_aoi = 1000, 
raster_template = my_template)

}

}
