% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordTable.R
\name{dplyr_reconstruct.records}
\alias{dplyr_reconstruct.records}
\title{Wrapper to discard class in dplyr}
\usage{
\method{dplyr_reconstruct}{records}(data, template)
}
\arguments{
\item{data}{A tibble. We use tibbles because they avoid some inconsistent
subset-assignment use cases.}

\item{template}{Template data frame to use for restoring attributes.}
}
\description{
This function is called internally by dplyr. It prevents the propagation of
attributes dependent on columns when such columns are filtered out.
}
\keyword{internal}
