% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_date.r
\name{exif_date}
\alias{exif_date}
\title{Extract EXIF DateTime from Jpeg}
\usage{
exif_date(img_path)
}
\arguments{
\item{img_path}{character; path to jpeg}
}
\value{
Character string representing the datetime as "Y:M:D H:M:S".
}
\description{
Extract the date and time a photo was taken from the it's EXIF data. This
funtion was heavily inspired by the R package \code{EXIFr}, and the
references listed below.
}
\examples{
img_path <- system.file("extdata", "muntjac.jpg", package = "camtrapr")
exif_date(img_path)
}
\references{
\url{http://code.flickr.net/2012/06/01/parsing-exif-client-side-using-javascript-2/}
   \url{https://github.com/cmartin/EXIFr/}
}

