% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.r
\name{plot.camOp}
\alias{plot.camOp}
\title{Plotting method for camera trap station operation matrix}
\usage{
\method{plot}{camOp}(x, camOp = NULL, palette = "viridis", lattice = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{camOp}{alternative to `x` (for backward compatibility)}

\item{palette}{the color palette to use (default: `"viridis"`)}

\item{lattice}{whether or not to plot using lattice (default: `FALSE`)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting method for camera trap station operation matrix
}
\keyword{internal}
