% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.r
\name{print.camOp}
\alias{print.camOp}
\title{Printing method for camera trap station operation matrix}
\usage{
\method{print}{camOp}(x, nRows = 4, nCols = 3, digits = 3, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{nRows}{the number of first and last rows to display}

\item{nCols}{the number of first and last columns to display}

\item{digits}{the number of digits to use within cells}

\item{...}{further arguments passed to or from other methods}
}
\description{
Printing method for camera trap station operation matrix
}
\keyword{internal}
