% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.r
\name{print.cams}
\alias{print.cams}
\title{Printing method for camera table objects}
\usage{
\method{print}{cams}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\description{
Printing method for camera table objects
}
\keyword{internal}
