% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.r
\name{print.cams_dp}
\alias{print.cams_dp}
\title{Printing method for Camtrap DP objects}
\usage{
\method{print}{cams_dp}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\description{
Printing method for Camtrap DP objects
}
\keyword{internal}
