% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.r
\name{summary.records}
\alias{summary.records}
\title{Summary method for record tables}
\usage{
\method{summary}{records}(object, nSpeciesMax = 5, ...)
}
\arguments{
\item{object}{an object of class \code{records}}

\item{nSpeciesMax}{integer for the maximum number of species to show in the
per-species detection table (defaults = 5)}

\item{...}{further arguments passed to or from other methods}
}
\value{
the species table (returned invisibly and unfiltered)
}
\description{
Summary method for record tables
}
\keyword{internal}
